Add a separate library for shared code
Ideally the macros should be exported through the egui_commonmark as
that would allow egui_commonmark to also reexport the code that the
generated code depends on. The code can't stay in egui_commonmark as
that would cause a cyclic dependency.
