+++ egui_commonmark/src/parsers/pulldown.rs
@@ -2,196 +2,13 @@
 
 use std::ops::Range;
 
-use crate::{elements::*, Alert, AlertBundle};
+use crate::elements::*;
 use crate::{CommonMarkCache, CommonMarkOptions};
 
-use egui::{self, Id, Pos2, TextStyle, Ui, Vec2};
+use egui::{self, Id, Pos2, TextStyle, Ui};
 
-use pulldown_cmark::{CowStr, HeadingLevel, Options};
-
-#[derive(Default, Debug)]
-pub struct ScrollableCache {
-    available_size: Vec2,
-    page_size: Option<Vec2>,
-    split_points: Vec<(usize, Pos2, Pos2)>,
-}
-
-pub type EventIteratorItem<'e> = (usize, (pulldown_cmark::Event<'e>, Range<usize>));
-
-/// Parse events until a desired end tag is reached or no more events are found.
-/// This is needed for multiple events that must be rendered inside a single widget
-fn delayed_events<'e>(
-    events: &mut impl Iterator<Item = EventIteratorItem<'e>>,
-    end_at: pulldown_cmark::TagEnd,
-) -> Vec<(pulldown_cmark::Event<'e>, Range<usize>)> {
-    let mut curr_event = events.next();
-    let mut total_events = Vec::new();
-    loop {
-        if let Some(event) = curr_event.take() {
-            total_events.push(event.1.clone());
-            if let (_, (pulldown_cmark::Event::End(tag), _range)) = event {
-                if end_at == tag {
-                    return total_events;
-                }
-            }
-        } else {
-            return total_events;
-        }
-
-        curr_event = events.next();
-    }
-}
-
-fn delayed_events_list_item<'e>(
-    events: &mut impl Iterator<Item = EventIteratorItem<'e>>,
-) -> Vec<(pulldown_cmark::Event<'e>, Range<usize>)> {
-    let mut curr_event = events.next();
-    let mut total_events = Vec::new();
-    loop {
-        if let Some(event) = curr_event.take() {
-            total_events.push(event.1.clone());
-            if let (_, (pulldown_cmark::Event::End(pulldown_cmark::TagEnd::Item), _range)) = event {
-                return total_events;
-            }
-
-            if let (_, (pulldown_cmark::Event::Start(pulldown_cmark::Tag::List(_)), _range)) = event
-            {
-                return total_events;
-            }
-        } else {
-            return total_events;
-        }
-
-        curr_event = events.next();
-    }
-}
-
-type Column<'e> = Vec<(pulldown_cmark::Event<'e>, Range<usize>)>;
-type Row<'e> = Vec<Column<'e>>;
-
-struct Table<'e> {
-    header: Row<'e>,
-    rows: Vec<Row<'e>>,
-}
-
-fn parse_row<'e>(
-    events: &mut impl Iterator<Item = (pulldown_cmark::Event<'e>, Range<usize>)>,
-) -> Vec<Column<'e>> {
-    let mut row = Vec::new();
-    let mut column = Vec::new();
-
-    for (e, src_span) in events.by_ref() {
-        if let pulldown_cmark::Event::End(pulldown_cmark::TagEnd::TableCell) = e {
-            row.push(column);
-            column = Vec::new();
-        }
-
-        if let pulldown_cmark::Event::End(pulldown_cmark::TagEnd::TableHead) = e {
-            break;
-        }
-
-        if let pulldown_cmark::Event::End(pulldown_cmark::TagEnd::TableRow) = e {
-            break;
-        }
-
-        column.push((e, src_span));
-    }
-
-    row
-}
-
-fn parse_table<'e>(events: &mut impl Iterator<Item = EventIteratorItem<'e>>) -> Table<'e> {
-    let mut all_events = delayed_events(events, pulldown_cmark::TagEnd::Table)
-        .into_iter()
-        .peekable();
-
-    let header = parse_row(&mut all_events);
-
-    let mut rows = Vec::new();
-    while all_events.peek().is_some() {
-        let row = parse_row(&mut all_events);
-        rows.push(row);
-    }
-
-    Table { header, rows }
-}
-
-/// try to parse events as an alert quote block. This ill modify the events
-/// to remove the parsed text that should not be rendered.
-/// Assumes that the first element is a Paragraph
-fn parse_alerts<'a>(
-    alerts: &'a AlertBundle,
-    events: &mut Vec<(pulldown_cmark::Event<'_>, Range<usize>)>,
-) -> Option<&'a Alert> {
-    // no point in parsing if there are no alerts to render
-    if !alerts.is_empty() {
-        let mut alert_ident = "".to_owned();
-        let mut alert_ident_ends_at = 0;
-        let mut has_extra_line = false;
-
-        for (i, (e, _src_span)) in events.iter().enumerate() {
-            if let pulldown_cmark::Event::End(_) = e {
-                // > [!TIP]
-                // >
-                // > Detect the first paragraph
-                // In this case the next text will be within a paragraph so it is better to remove
-                // the entire paragraph
-                alert_ident_ends_at = i;
-                has_extra_line = true;
-                break;
-            }
-
-            if let pulldown_cmark::Event::SoftBreak = e {
-                // > [!NOTE]
-                // > this is valid and will produce a soft break
-                alert_ident_ends_at = i;
-                break;
-            }
-
-            if let pulldown_cmark::Event::HardBreak = e {
-                // > [!NOTE]<whitespace>
-                // > this is valid and will produce a hard break
-                alert_ident_ends_at = i;
-                break;
-            }
-
-            if let pulldown_cmark::Event::Text(text) = e {
-                alert_ident += text;
-            }
-        }
-
-        let alert = alerts.try_get_alert(&alert_ident);
-
-        if alert.is_some() {
-            // remove the text that identifies it as an alert so that it won't end up in the
-            // render
-            //
-            // FIMXE: performance improvement potential
-            if has_extra_line {
-                for _ in 0..=alert_ident_ends_at {
-                    events.remove(0);
-                }
-            } else {
-                for _ in 0..alert_ident_ends_at {
-                    // the first element must be kept as it _should_ be Paragraph
-                    events.remove(1);
-                }
-            }
-        }
-
-        alert
-    } else {
-        None
-    }
-}
-
-/// Supported pulldown_cmark options
-fn parser_options() -> Options {
-    Options::ENABLE_TABLES
-        | Options::ENABLE_TASKLISTS
-        | Options::ENABLE_STRIKETHROUGH
-        | Options::ENABLE_FOOTNOTES
-}
+use egui_commonmark_shared::pulldown::*;
+use pulldown_cmark::{CowStr, HeadingLevel};
 
 pub struct CommonMarkViewerInternal {
     source_id: Id,
