+++ egui_commonmark_macro/tests/fail/commonmark_str_not_found.rs
@@ -1,0 +1,10 @@
+use egui::__run_test_ui;
+use egui_commonmark_macro::commonmark_str;
+
+// Check that it fails to compile when it is not able to find a file
+fn main() {
+    let mut cache = egui_commonmark_shared::CommonMarkCache::default();
+    __run_test_ui(|ui| {
+        commonmark_str!("a", ui, &mut cache, "foo.md");
+    });
+}
