+++ egui_commonmark_macro/tests/pass/ui_hygiene.rs
@@ -1,0 +1,14 @@
+use egui::__run_test_ui;
+use egui_commonmark_macro::commonmark;
+
+// Check hygiene of the ui expression
+fn main() {
+    let mut cache = egui_commonmark_shared::CommonMarkCache::default();
+    __run_test_ui(|ui| {
+        egui::ScrollArea::vertical().show(ui, |ui| {
+            egui::Frame::none().show(ui, |not_named_ui| {
+                commonmark!("a", not_named_ui, &mut cache, "# Hello, World");
+            })
+        });
+    });
+}
