+++ egui_commonmark_shared/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "egui_commonmark_shared"
+authors = ["Erlend Walstad"]
+version = "0.15.0"
+edition = "2021"
+description = "Shared code for egui_commonmark and egui_commonmark_macro"
+keywords = ["commonmark", "egui"]
+categories = ["gui"]
+license = "MIT OR Apache-2.0"
+readme = "README.md"
+repository = "https://github.com/lampsitter/egui_commonmark"
+documentation = "https://docs.rs/egui_commonmark"
+include = ["**/*.rs", "LICENSE-MIT", "LICENSE-APACHE", "Cargo.toml"]
+rust-version = "1.76" # Required by egui
+
+[dependencies]
+pulldown-cmark = { version = "0.10", default-features = false, optional = true }
+egui_extras = { workspace = true }
+egui = { workspace = true }
+
+
+syntect = { version = "5.0.0", optional = true, default-features = false, features = [
+    "default-fancy",
+] }
+
+[features]
+better_syntax_highlighting = ["dep:syntect"]
+pulldown-cmark = ["dep:pulldown-cmark"]
